Graph locking now works as intended.

+++ backend/src/main/java/io/dcbn/backend/graph/controllers/GraphController.java
@@ -2,6 +2,7 @@ package io.dcbn.backend.graph.controllers;
 
 import io.dcbn.backend.graph.Graph;
 import io.dcbn.backend.graph.Node;
+import io.dcbn.backend.graph.PotentiallyLockedGraph;
 import io.dcbn.backend.graph.ValueNode;
 import io.dcbn.backend.graph.converters.GenieConverter;
 import io.dcbn.backend.graph.repositories.GraphRepository;
@@ -11,8 +12,10 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.request.async.DeferredResult;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 import org.xml.sax.SAXException;
@@ -23,18 +26,25 @@ import javax.xml.transform.TransformerException;
 import java.io.IOException;
 import java.security.Principal;
 import java.util.*;
-import java.util.function.Supplier;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 @RestController
-@PreAuthorize("hasRole('ADMIN')")
 public class GraphController {
 
     private final GraphRepository repository;
     private final GraphService graphService;
     private static final String GRAPH_NOT_FOUND = "Graph not found";
+
+    private final Map<String, DeferredResult<Iterable<PotentiallyLockedGraph>>> resultMap = new ConcurrentHashMap<>();
+
+    @Scheduled(fixedRate = 3000)
+    public void updateLocks() {
+        if (graphService.updateExpiredLocks()) {
+            updateResults();
+        }
+    }
 
     @Autowired
     public GraphController(GraphRepository repository, GraphService graphService) {
@@ -42,42 +52,47 @@ public class GraphController {
         this.graphService = graphService;
     }
 
+    private Iterable<PotentiallyLockedGraph> getPotentiallyLockedGraphs(String username) {
+        return StreamSupport.stream(repository.findAll().spliterator(), false)
+                .peek(graph -> graph.setNodes(Collections.emptyList()))
+                .map(graph -> new PotentiallyLockedGraph(graph, graphService.cannotAccess(graph.getId(), username)))
+                .collect(Collectors.toList());
+    }
+
+    private void updateResults() {
+        for (Map.Entry<String, DeferredResult<Iterable<PotentiallyLockedGraph>>> result : resultMap.entrySet()) {
+            result.getValue().setResult(getPotentiallyLockedGraphs(result.getKey()));
+        }
+        resultMap.clear();
+    }
+
     @GetMapping("/graphs")
     @PreAuthorize("hasRole('MODERATOR')")
-    public Iterable<Graph> getGraphs(@RequestParam boolean withStructure) {
-        Supplier<Stream<Graph>> streamSupplier = () -> StreamSupport.stream(repository.findAll().spliterator(), false);
-        streamSupplier.get().forEach(graph -> {
-            if (!withStructure) {
-                graph.setNodes(Collections.emptyList());
-            }
-        });
-        return streamSupplier.get().collect(Collectors.toList());
+    public DeferredResult<Iterable<PotentiallyLockedGraph>> getGraphs(@RequestParam(required = false, defaultValue = "false") boolean delayed, Principal principal) {
+        DeferredResult<Iterable<PotentiallyLockedGraph>> result =
+                new DeferredResult<>(60000L, getPotentiallyLockedGraphs(principal.getName()));
+
+        if (!delayed) {
+            result.setResult(getPotentiallyLockedGraphs(principal.getName()));
+        } else {
+            resultMap.put(principal.getName(), result);
+        }
+        return result;
     }
 
     @GetMapping("/graphs/{id}")
     @PreAuthorize("hasRole('MODERATOR')")
     public Graph getGraphById(@PathVariable long id, Principal principal) {
-        Graph graph = repository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
-        // TODO: Currently only one user can access the graph structure,
-        //  it doesn't matter whether the user is testing or editing the graph.
-        try {
-            graphService.updateLock(graph.getId(), principal.getName());
-        } catch (IllegalArgumentException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        if (graphService.cannotAccess(id, principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
-        return graph;
+        return repository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     @PostMapping("/graphs")
-    public long createGraph(@Valid @RequestBody Graph graph, Principal principal) {
-        graph = repository.save(graph);
-        try {
-            graphService.updateLock(graph.getId(), principal.getName());
-        } catch (IllegalArgumentException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
-        }
-
+    @PreAuthorize("hasRole('ADMIN')")
+    public long createGraph(@Valid @RequestBody Graph graph) {
         if (graphService.hasCycles(graph)) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Graph has cycles!");
         }
@@ -86,32 +101,31 @@ public class GraphController {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Some evidence formulas do not exist!");
         }
 
-        return graph.getId();
+        long id = repository.save(graph).getId();
+        updateResults();
+        return id;
     }
 
     @DeleteMapping("/graphs/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public void deleteById(@PathVariable long id, Principal principal) {
+        if (graphService.cannotAccess(id, principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
         if (!repository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, GRAPH_NOT_FOUND);
         }
 
-        try {
-            graphService.updateLock(id, principal.getName());
-        } catch (IllegalArgumentException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
-        }
-
         repository.deleteById(id);
+        updateResults();
     }
 
     @PutMapping("/graphs/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public void updateGraphById(@PathVariable long id, @Valid @RequestBody Graph graph, Principal principal) {
-        try {
-            graphService.updateLock(graph.getId(), principal.getName());
-        } catch (IllegalArgumentException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+        if (graphService.cannotAccess(id, principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
-
         if (graphService.hasCycles(graph)) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Graph has cycles!");
         }
@@ -126,10 +140,12 @@ public class GraphController {
         oldGraph.getNodes().addAll(graph.getNodes());
         oldGraph.setName(graph.getName());
         oldGraph.setTimeSlices(graph.getTimeSlices());
+
         repository.save(oldGraph);
     }
 
     @PostMapping("/graphs/{id}/name")
+    @PreAuthorize("hasRole('ADMIN')")
     public void renameGraphById(@PathVariable long id, @RequestBody String name) {
         Graph graph = repository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, GRAPH_NOT_FOUND));
         //Checking that the graph name is unique
@@ -142,6 +158,7 @@ public class GraphController {
 
         graph.setName(name);
         repository.save(graph);
+        updateResults();
     }
 
     @PostMapping("/graphs/{id}/evaluate")
@@ -186,9 +203,12 @@ public class GraphController {
     }
 
     @PutMapping("/graphs/{id}/lock")
+    @PreAuthorize("hasRole('MODERATOR')")
     public void updateGraphLockById(@PathVariable long id, Principal principal) {
         try {
-            graphService.updateLock(id, principal.getName());
+            if (graphService.updateLock(id, principal.getName())) {
+                updateResults();
+            };
         } catch (IllegalArgumentException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
         }
@@ -196,6 +216,7 @@ public class GraphController {
 
 
     @PostMapping("/graphs/import")
+    @PreAuthorize("hasRole('ADMIN')")
     public Graph importGraphFromGenie(@RequestParam("graph") MultipartFile uploadedFile) {
         if (uploadedFile.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No file uploaded");
@@ -218,9 +239,11 @@ public class GraphController {
             }
         }
         repository.save(graph);
-        return repository.findById(graph.getId())
+        graph = repository.findById(graph.getId())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Save Failed"));
 
+        updateResults();
+        return graph;
     }
 
     @GetMapping("/graphs/{id}/export")
